Fix so that we can upload clips on the camera
This fix is not yet merged into master. It was created by hassanctech on the AWS team to help Owlet get
passed a deadlock bug that would cause our frame upload thread to lock.

Once the fix has been merged into the official repository this commit on the Owlet fork will no longer
be necessary.